Change type from int to uint64 in getTransactionCount

+++ ethrpc.go
@@ -249,14 +249,14 @@ func (rpc *EthRPC) EthGetStorageAt(data string, position int, tag string) (strin
 }
 
 // EthGetTransactionCount returns the number of transactions sent from an address.
-func (rpc *EthRPC) EthGetTransactionCount(address, block string) (int, error) {
+func (rpc *EthRPC) EthGetTransactionCount(address, block string) (uint64, error) {
 	var response string
 
 	if err := rpc.call("eth_getTransactionCount", &response, address, block); err != nil {
 		return 0, err
 	}
 
-	return ParseInt(response)
+	return ParseUint64(response)
 }
 
 // EthGetBlockTransactionCountByHash returns the number of transactions in a block from a block matching the given block hash.
